requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
entrypoints==0.4
fastapi==0.95.0
ffmpy==0.3.0
filelock==3.10.7
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
gradio==3.24.1
gradio_client==0.0.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
orjson==3.8.9
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pydantic==1.10.7
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
websockets==11.0
yarl==1.8.2
